consolidate integration tests
